- Refactored CLI into separate files per command. - Added initial version of 'walltowall clone' command: CBM4-only, creates a new CBM4 project based on an existing one and sets up the cache configuration for the new run to continue from the old one. - Spatial rollback random seed can now be specified for consistent distribution draws.
