+++ gcbmwalltowall/application/command/argbase.py
@@ -1,0 +1,7 @@
+class ArgBase(dict):
+
+    def __getattr__(self, key):
+        return self[key]
+
+    def __setattr__(self, key, value):
+        self[key] = value
