+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "gcbmwalltowall"
-version = "2.11.3"
+version = "2.12.0"
 description = "gcbmwalltowall"
 classifiers = ["Programming Language :: Python :: 3"]
 requires-python = ">=3.9"
